requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
openai==1.3.0
python-dotenv==1.0.0
numpy==1.25.2
pandas==2.1.0
httpx==0.24.1
python-multipart==0.0.6
